+++ doc/changes.rst
@@ -16,12 +16,14 @@ Other
 * Add "UPDATED" timestamp column to tiles file (PR `#2373`_).
 * Support reading old zbest files (now redrock) (PR `#2374`_).
 * Add script to update processing tabel column layout (PR `#2376`_).
+* Fix column incompatibility in :func:`desispec.specstatus.update_specstatus` (PR `#2384`_).
 
 .. _`#2365`: https://github.com/desihub/desispec/pull/2365
 .. _`#2373`: https://github.com/desihub/desispec/pull/2373
 .. _`#2374`: https://github.com/desihub/desispec/pull/2374
 .. _`#2376`: https://github.com/desihub/desispec/pull/2376
 .. _`#2377`: https://github.com/desihub/desispec/pull/2377
+.. _`#2384`: https://github.com/desihub/desispec/pull/2384
 
 0.66.3 (2024-09-13)
 -------------------